"deeplabcut.filterpredictions" and "deeplabcut.analyze_skeleton" return mappings (#1908)
* REF: Denest filterpredictions for loop

there were two try/excepts, one within the other.
this commit simplifies the code by pulling the inner try/except out.

the code now tries to look for filtered data. if filtered data is found
for the video, we continue to the next item in the loop.

if not found, we catch and ignore the error. we then try to load the
data and filter it

	modified:   deeplabcut/post_processing/filtering.py

* REF: Reduce the code within the try block

the second try block in filterpredictions looked for the analyzed file
and performed operations to filter it - and if the file is not found,
continue to the next item in the for loop.

this commit reduces the try block to only load the analyzed file and
immediately catch the error and continue if the file doesnt exist.

the rest of the filtering on the analyzed data is performed outside the
try/except block.

	modified:   deeplabcut/post_processing/filtering.py

* FEAT: Return video filename to filtered dataframe mapping

the filterpredictions now returns a mapping instead of returning None.

	modified:   deeplabcut/post_processing/filtering.py

* REF: Reduce scope of try/except

instead of finding and analyzing the skeleton in the try block, this
commit reduces the scope of the try block to only finding the file and
moves the skeleton calculations to outside the try/except.

	modified:   deeplabcut/post_processing/analyze_skeleton.py

* FIX: Actually return the mapping

	modified:   deeplabcut/post_processing/filtering.py

* FEAT: Return mapping from video filename to skeleton dataframe

the analyze_skeleton function now returns a mapping instead of None

	modified:   deeplabcut/post_processing/analyze_skeleton.py

* Add return_data flag

---------

Co-authored-by: Jessy Lauer <30733203+jeylau@users.noreply.github.com>